python_backend/requirements.txt
flask
opencv-python
face_recognition
numpy
supabase
python-dotenv
requests
